+++ bin/langfuse-ecr-ecs-deployment-cdk.ts
@@ -3,10 +3,9 @@ import 'source-map-support/register';
 
 import * as cdk from 'aws-cdk-lib';
 import * as dotenv from 'dotenv';
-import { CdkLangfuseEcrEcsDeploymentStack } from '../lib/langfuse-ecr-ecs-deployment-cdk-stack';
+import { CdkLangfuseEcrEcsFargateDeploymentStack } from '../lib/langfuse-ecr-ecs-fargate-deployment-cdk-stack';
 import { checkEnvVariables } from '../utils/check-environment-variable';
 import { IEnvTypes } from '../process-env-typed';
-import { CdkLangfuseEcrEcsDeploymentStack2 } from '../lib/langfuse-ecr-ecs-deployment-cdk-stack2';
 
 dotenv.config(); // Load environment variables from .env file
 const app = new cdk.App();
@@ -21,49 +20,25 @@ export const LATEST_IMAGE_VERSION = 'latest';
 // check general stack props
 checkEnvVariables('ECR_REPOSITORY_NAME', 'APP_NAME', 'IMAGE_VERSION', 'PORT');
 
-const envTyped: IEnvTypes = {
-    ECR_REPOSITORY_NAME: process.env.ECR_REPOSITORY_NAME!,
-    APP_NAME: process.env.APP_NAME!,
-    IMAGE_VERSION: process.env.IMAGE_VERSION!,
-    PORT: process.env.PORT!,
-};
+const appName = process.env.APP_NAME!;
 
 for (const cdkRegion of cdkRegions) {
   for (const environment of deployEnvironments) {
-    new CdkLangfuseEcrEcsDeploymentStack(app, `${envTyped.APP_NAME}-${environment}-${cdkRegion}-CdkLangfuseEcrEcsDeploymentStack`, {
+    new CdkLangfuseEcrEcsFargateDeploymentStack(app, `${appName}-${environment}-${cdkRegion}-CdkLangfuseEcrEcsDeploymentStack`, {
       env: {
         account,
         region: cdkRegion,
       },
       tags: {
         environment,
-        appName: envTyped.APP_NAME,
+        appName: appName,
         AppManagerCFNStackKey: 'true',
       },
       deployRegion: cdkRegion,
       environment,
-      envTyped,
-      appName: envTyped.APP_NAME,
-      stackName: `${envTyped.APP_NAME}-${environment}-${cdkRegion}-CdkLangfuseEcrEcsDeploymentStack`,
+      appName,
+      stackName: `${appName}-${environment}-${cdkRegion}-CdkLangfuseEcrEcsDeploymentStack`,
       description: `Langfuse ECR/ECS deployment stack for ${environment} environment in ${cdkRegion} region.`,
     });
-
-    // new CdkLangfuseEcrEcsDeploymentStack2(app, `${envTyped.APP_NAME}-${environment}-${cdkRegion}-CdkLangfuseEcrEcsDeploymentStack2`, {
-    //   env: {
-    //     account,
-    //     region: cdkRegion,
-    //   },
-    //   tags: {
-    //     environment,
-    //     appName: envTyped.APP_NAME,
-    //     AppManagerCFNStackKey: 'true',
-    //   },
-    //   deployRegion: cdkRegion,
-    //   environment,
-    //   envTyped,
-    //   appName: envTyped.APP_NAME,
-    //   stackName: `${envTyped.APP_NAME}-${environment}-${cdkRegion}-CdkLangfuseEcrEcsDeploymentStack2`,
-    //   description: `Langfuse ECR/ECS deployment stack for ${environment} environment in ${cdkRegion} region.`,
-    // });
   }
 }
