--- lib/LangfuseAppRunnerWithVpcDeploymentStackProps.ts
@@ -1,7 +1,0 @@
-import * as cdk from "aws-cdk-lib";
-import { LangfuseDockerImageEcsDeploymentCdkStackProps } from "./LangfuseDockerImageEcsDeploymentCdkStackProps";
-
-export interface CdkAppRunnerWithVpcDeploymentStackProps extends LangfuseDockerImageEcsDeploymentCdkStackProps{
-    readonly dbServerSG: cdk.aws_ec2.SecurityGroup;
-    readonly vpcId: string;
-}
