+++ lib/LangfuseEcsStackProps.ts
@@ -1,9 +1,9 @@
 import * as cdk from 'aws-cdk-lib';
 import * as ec2 from 'aws-cdk-lib/aws-ec2';
-import { DatabaseEnvTyped, DockerEnvTyped } from '../process-env-typed';
-import { LangfuseDockerImageEcrDeploymentCdkStackProps } from './LangfuseDockerImageEcrDeploymentCdkStackProps';
+import { DockerEnvTyped } from '../process-env-typed';
+import { LangfuseEcrStackProps } from './LangfuseEcrStackProps';
 
-export interface LangfuseDockerImageEcsDeploymentCdkStackProps extends LangfuseDockerImageEcrDeploymentCdkStackProps {
+export interface LangfuseEcsStackProps extends LangfuseEcrStackProps {
     /**
      * The port number on which the container service will be available.
      */
@@ -20,4 +20,6 @@ export interface LangfuseDockerImageEcsDeploymentCdkStackProps extends LangfuseD
      * The VPC where the ECS services and other resources will be deployed.
      */
     readonly vpc: ec2.Vpc;
+
+    readonly dbServerSG: cdk.aws_ec2.SecurityGroup;
 }
