+++ lib/LangfusePostgresStackProps.ts
@@ -1,8 +1,8 @@
 import * as ec2 from 'aws-cdk-lib/aws-ec2';
 import { DatabaseEnvTyped } from '../process-env-typed';
-import { LangfuseDockerImageEcrDeploymentCdkStackProps } from './LangfuseDockerImageEcrDeploymentCdkStackProps';
+import { LangfuseEcrStackProps } from './LangfuseEcrStackProps';
 
-export interface LangfusePostgresStackProps extends LangfuseDockerImageEcrDeploymentCdkStackProps {
+export interface LangfusePostgresStackProps extends LangfuseEcrStackProps {
     /**
      * The database environment variables and arguments.
      */
